Add files via upload
Data Visualization is a must-skill for  data science. The goal is to understand the data relations and to select features that are useful for our machine Learning problems.
There are 2 data visualization techniques for this tutorial (1) Bar Plot and (2)Histogram.
But u need to note that there are many other way to do that.